% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.kappalate.R
\name{print.kappalate}
\alias{print.kappalate}
\title{Print kappalate Results}
\usage{
\method{print}{kappalate}(x, ...)
}
\arguments{
\item{x}{A kappalate object}

\item{...}{Additional parameters (not used)}
}
\description{
Prints a kappalate Object (formatted like Stata)
}
